Simplify Spring Boot configuration
This commit removes some duplication in the atsea application.

The `pom.xml` file has been cleaned to reflect the fact that:

* some dependencies weren't used
* others were already transitive dependencies
* and/or their version was already managed by Boot's BOM

The Configuration classes were also refactored:

* in `AtSeaApp`, removed annotations were either implicit or the
Boot auto-configuration already does the job
* `BeanConfiguration` is not required, as components are already
declared as such with `@Service` annotations
* `JpaConfiguration` has been removed, as most customizations
can be applied through configuration keys in `application.yml`
and managing the secret password can be done with an
`EnvironmentPostProcessor` (more on that later)
* the web configuration does not need `@EnableWebMvc` as this
completely overrides the Boot auto-configuration for the web parts
* furthermore, the static locations can be configured though properties

This commit adds an `EnvironmentPostProcessor`, which updates
the configuration key/values in the application environment as the
application starts up. Once the postgresql password is added, the
other datasource properties and overrides aren't needed anymore.

Note: most configuration keys related to datasource configuration aren't
applicable to hikari. We can reinstate the previous values with Tomcat
DBCP (which should work with those) or leave this as it is.